Make buckets of PCRs
Instead of having all PCRs in a single linked list, make a linked list
of buckets of PCRs. Each element of the outer linked list has an
associated SCC. Each individual bucket is just a linked list of
objects, same as the PCR struct before. This new structure means you
don't need to traverse a bunch of objects to insert a new PCR - you just
need to go over a handful of buckets until you get to the right SCC.