fix: Fix use after free with SetFieldExprso
Previously, it was decrementing the old value before incrementing the
new value. This could be problematic if the old value got freed but the
new value actually had a reference to the old value. For example,
foo.list = Cons { head: "new", tail: foo.list } (for appending to a
list) would cause a use after free.